requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
python-pptx>=0.6.23
Pillow>=10.0.0
